% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_df_sce}
\alias{cell_type_df_sce}
\title{cell_type_df_sce}
\format{
A tibble containing two columns: label and first.labels
}
\usage{
data(cell_type_df_sce)
}
\description{
Stefano's cell type df for tidySingleCellExperiment
}
\keyword{datasets}
