% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_df_seurat}
\alias{cell_type_df_seurat}
\title{cell_type_df_seurat}
\format{
A tibble containing two columns: seurat_clusters and first.labels
}
\usage{
data(cell_type_df_seurat)
}
\description{
Stefano's cell type df for tidyseurat
}
\keyword{datasets}
