% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small_sce}
\alias{pbmc_small_sce}
\title{pbmc_small_sce}
\format{
A SingleCellExperiment object.
}
\usage{
data(pbmc_small_sce)
}
\description{
A small SingleCellExperiment dataset of single cell RNA sequencing data
}
\keyword{datasets}
