% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small_seurat}
\alias{pbmc_small_seurat}
\title{pbmc_small_seurat}
\format{
A Seurat object.
}
\usage{
data(pbmc_small_seurat)
}
\description{
A small Seurat dataset of single cell RNA sequencing data
}
\keyword{datasets}
